function changed to float

+++ Determinant.cpp
@@ -1,39 +1,39 @@
 #include<iostream>
 using namespace std;
 int main(){
-    int a11;
+    float a11;
     cout<<"a11 : ";
     cin>>a11;
 
-    int a12;
+    float a12;
     cout<<"a12 : ";
     cin>>a12;
 
-    int a13;
+    float a13;
     cout<<"a13 : ";
     cin>>a13;
 
-    int a21;
+    float a21;
     cout<<"a21 : ";
     cin>>a21;
 
-    int a22;
+    float a22;
     cout<<"a22 : ";
     cin>>a22;
 
-    int a23;
+    float a23;
     cout<<"a23 : ";
     cin>>a23;
 
-    int a31;
+    float a31;
     cout<<"a31 : ";
     cin>>a31;
 
-    int a32;
+    float a32;
     cout<<"a32 : ";
     cin>>a32;
 
-    int a33;
+    float a33;
     cout<<"a33 : ";
     cin>>a33;
